Implemented DuDe smell detector. Fixed an issue with duplicate smells.
The Set of smells is changed to a Map with a Set for each type of smell.
Also, the hashCode() and equals() method is overriden in each smell
class in order to define it for each one correctly and find duplicates
when added to the Set of the corresponding smell.